Search for wp-config.php above document root.
git-svn-id: http://mu.svn.wordpress.org/trunk@1407 7be80a69-a1ef-0310-a953-fb0f7c49ff36

+++ wp-blog-header.php
@@ -6,7 +6,7 @@
  */
 
 if ( !isset($wp_did_header) ) {
-	if ( !file_exists( dirname(__FILE__) . '/wp-config.php') ) {
+	if ( !file_exists( dirname(__FILE__) . '/wp-config.php') && !file_exists( dirname( dirname(__FILE__) ) . '/wp-config.php')) {
 		if (strpos($_SERVER['PHP_SELF'], 'wp-admin') !== false) $path = '';
 		else $path = 'wp-admin/';
 		include( "index-install.php" ); // install WPMU!
